thermal: Fix ABI change in 4.4.4
struct thermal_zone_device is always allocated in thermal_core and the
new field is only used there.  So we can safely move it to the end of the
structure and hide it from genksyms.

Also hide the new #include, as if it could make an actual difference,
stupid genksyms...


Gbp-Pq: Topic debian
Gbp-Pq: Name thermal-fix-abi-change-in-4.4.4.patch